database and schema creation
Rough draft of the database file, and creation of the schema file. The schema file will start a local server of MongoDB running on port 27017, and will create the default admin user. The database file will only work on my end until I connect it to some sort of domain or authorize users.